+++ test/rusty_opus/reencode_test.exs
@@ -21,10 +21,32 @@ defmodule RustyOpus.ReencodeTest do
     assert div(us, 1000) < @max_reencode_ms
   end
 
-  test "reencode errors" do
+  test "voip and audio applications produce different bitstreams" do
+    source = File.read!(@fixture)
+
+    assert {:ok, voip} = RustyOpus.reencode(source, bitrate: 20_000, application: :voip)
+
+    assert {:ok, audio} =
+             RustyOpus.reencode(source, bitrate: 20_000, application: :audio, complexity: 10)
+
+    assert byte_size(voip) > 0
+    assert voip != audio
+  end
+
+  test "reencode rejects unsupported frame durations and bad options" do
     source = File.read!(@fixture)
     assert {:error, %Error{reason: :invalid_settings}} = RustyOpus.reencode(source, [])
     assert {:error, %Error{reason: :invalid_settings}} = RustyOpus.reencode(source, bitrate: 0)
+
+    assert {:error, %Error{reason: :invalid_settings}} =
+             RustyOpus.reencode(source, bitrate: 20_000, frame_duration_ms: 60)
+
+    assert {:error, %Error{reason: :invalid_settings}} =
+             RustyOpus.reencode(source, bitrate: 20_000, complexity: 99)
+
+    assert {:error, %Error{reason: :invalid_settings}} =
+             RustyOpus.reencode(source, bitrate: 20_000, application: :music)
+
     assert {:error, %Error{reason: :invalid_input}} = RustyOpus.reencode(<<>>, bitrate: 20_000)
 
     assert {:error, %Error{reason: :decode_failed}} =